XJ - moved all java files about child register to its own folder
