+++ dristhi-app/registers/child/src/main/java/org/ei/drishti/view/activity/NativeChildSmartRegisterActivity.java
