+++ dristhi-app/registers/child/src/main/java/org/ei/drishti/view/viewHolder/ChildRegisterProfilePhotoLoader.java
